requirements.txt
fastapi
uvicorn
pymongo
pytz
dnspython
fastapi-login
python-multipart
blurhash-python
boto3
Pillow